+++ Integrations/SplunkPy/CHANGELOG.md
@@ -1,4 +1,25 @@
 ## [Unreleased]
+Removed newlines from Markdown headers to facilitate converting Markdown entries to HTML.
+
+## [20.1.0] - 2020-01-07
+Fixed an issue with the access to a non-existing key when fetching non-ES events.
+
+## [19.12.1] - 2019-12-25
+Enhanced the execution speed of the ***splunk-search*** command.
+
+## [19.11.0] - 2019-11-12
+Increased the maximum fetch limit for Splunk.
+
+## [19.10.2] - 2019-10-29
+  - Improved handling of the *app context* parameter.
+  - Fixed handling of arrays when converting notable events to incidents.
+
+## [19.10.1] - 2019-10-15
+- Added the *app* parameter, which is the app context of the namespace.
+- Prettified the human readable of the search command.
+
+
+## [19.10.0] - 2019-10-03
 Added the *Earliest time to fetch* and *Latest time to fetch* parameters, which are the name of the Splunk fields whose value defines the query's earliest and latest time to fetch.